workers/requirements.txt
ultralytics
opencv-python-headless
boto3
redis
pymongo
python-dotenv
filterpy
scipy
scikit-image
lap
